M2M-021_Commit02_Phat: Done

+++ clothing/src/main/java/m2m_phase2/clothing/clothing/data/entity/PaymentE.java
@@ -24,7 +24,7 @@ public class PaymentE {
     private String payerId;
 
     @Column(name = "total_amount", nullable = false)
-    private double totalAmount;
+    private Float totalAmount;
 
     @Column(name = "currency", columnDefinition = "nvarchar(10) default 'USD'")
     private String currency = "USD";
